refactor: 기존 monthlyData.ts에 있던 데이터 페칭 관련 함수 부분을 각각의 hook로 호출하는 방식으로 전체 변경
